Move the ODT parsing code into its own file
